Be more specific about what expand-polygon can do
As discussed in horizon-eda/horizon-pool-convention#13, it's complicated and I made some incorrect assumptions after reading this. I think making it clear here will reach anybody who goes looking.

The question of how much detail to specify in horizon-pool-convention seems separate to me, because the docs for the parameter program commands should be clear on their own.